Password Hashing.
I have implemented password hashing into the solution to improve the application's security. This is an extremely important feature of the solution as my sensitive data will be stored within the database. By making it more difficult for hackers to obtain any users data, the likelihood of a data breach reduces.