Protected UI Pages !!! Finally!!
I have now implemented functionaing cookies and session into the applciation, which will allow different pages to only be accessed once the user has logged in at the account management module. When the user attempts to log in, an express-session will be created, which contain the Clients object ID (MongoDb generated) and the users email. This will also be within the cookie, but encrypted via the secrete code within the Env file.

This element will allow me to create user specific content, such as graph generation for unique businesses.